add context to resolvers, include gameId

+++ project-3-app/server/schemas/resolvers.js
@@ -73,35 +73,60 @@ const resolvers = {
 
         // THE USER YOU CHOOSE TO GENERATE YOUR TOKEN ON THE LOGIN ROUTE WILL BE THE USER AFFECTED BY ALL AUTH ROUTES, SO IT DOESN'T MATTER WHAT "AUTHOR" ID IS USED IN THE APOLLO VARIABLES
 
-        addPost: async (parent, { postTitle, postText }, context) => {
+        addPost: async (parent, { postTitle, postText, gameId }, context) => {
             console.log("test addPost");
+            if (!context.user) {
+                throw new AuthenticationError('You need to be logged in to add a post');
+            }
             try {
-              if (context.user) {
                 const post = await Post.create({
-                  postTitle,
-                  postText,
-                  postAuthor: context.user._id,
+                    postTitle,
+                    postText,
+                    postAuthor: context.user._id,
+                    game: gameId // or whatever we use to identify the games through the API
                 });
                 console.log("author user id is: " + context.user._id)
-          
+
                 await User.findOneAndUpdate(
-                  { _id: context.user._id },
-                  { $addToSet: { posts: post._id } }
+                    { _id: context.user._id },
+                    { $addToSet: { posts: post._id } }
                 );
-                console.log("user id is: " + context.user._id )
+                console.log("user id is: " + context.user._id)
                 return post;
-              }
+
             } catch (error) {
-              throw new Error(`Error creating post: ${error.message}`);
+                throw new Error(`Error creating post: ${error.message}`);
             }
         },
 
 
-        updatePost: async (parent, { postId, postTitle, postText }) => {
+        updatePost: async (parent, { postId, postTitle, postText, gameId }, context) => {
+            console.log("test updatePost");
+            if (!context.user) {
+                throw new AuthenticationError('You need to be logged in to update a post');
+            }
             try {
+                const existingPost = await Post.findOne({
+                    _id: postId,
+                    postAuthor: context.user._id
+                });
+                if (!existingPost) {
+                    throw new AuthenticationError('You are not authorized to update this post');
+                }
+
+                const updates = {
+                    postTitle,
+                    postText,
+                };
+        
+                // if gameId is provided, update the game association
+                if (gameId) {
+                    updates.game = gameId;
+                }
+
                 const updatedPost = await Post.findOneAndUpdate(
                     { _id: postId },
-                    { $set: { postTitle, postText } },
+                    { $set: updates },
                     { new: true }
                 );
 
@@ -114,8 +139,21 @@ const resolvers = {
                 throw new Error(`Error updating post: ${error.message}`);
             }
         },
-        removePost: async (parent, { postId }) => {
+
+        removePost: async (parent, { postId }, context) => {
+            console.log("test removePost");
+            if (!context.user) {
+                throw new AuthenticationError('You need to be logged in to remove a post');
+            }
             try {
+                const existingPost = await Post.findOne({
+                    _id: postId,
+                    postAuthor: context.user._id
+                });
+
+                if (!existingPost) {
+                    throw new AuthenticationError('You are not authorized to remove this post');
+                }
                 const deletedPost = await Post.findOneAndDelete({ _id: postId });
 
                 if (!deletedPost) {
@@ -129,10 +167,12 @@ const resolvers = {
                 );
 
                 // remove the reference to this post from the game
-                await Game.findOneAndUpdate(
-                    { _id: deletedPost.game },
-                    { $pull: { posts: postId } }
-                );
+                if (deletedPost.game) {
+                    await Game.findOneAndUpdate(
+                        { _id: deletedPost.game },
+                        { $pull: { posts: postId } }
+                    );
+                }
 
                 // do we want to delete all comments associated with a Post, or make a placeholder "user removed post" that leaves comments visible under the deleted post?
                 // await Comment.deleteMany({ post: postId });
@@ -143,12 +183,21 @@ const resolvers = {
             }
         },
 
-        addComment: async (parent, { postId, commentText, commentAuthor }) => {
+        addComment: async (parent, { postId, commentText, commentAuthor }, context) => {
+            console.log("test addComment");
+            if (!context.user) {
+                throw new AuthenticationError('You need to be logged in to comment on other posts');
+            }
             try {
                 const updatedPost = await Post.findOneAndUpdate(
                     { _id: postId },
                     {
-                        $addToSet: { comments: { commentText, commentAuthor } },
+                        $addToSet: {
+                            comments: {
+                                commentText,
+                                commentAuthor: context.user._id
+                            }
+                        },
                     },
                     {
                         new: true,
@@ -165,10 +214,17 @@ const resolvers = {
                 throw new Error(`Error adding comment: ${error.message}`);
             }
         },
-        updateComment: async (parent, { commentId, commentText }) => {
+        updateComment: async (parent, { commentId, commentText }, context) => {
+            console.log("test updateComment");
+            if (!context.user) {
+                throw new AuthenticationError('You need to be logged in to update a comment');
+            }
             try {
                 const updatedComment = await Comment.findOneAndUpdate(
-                    { _id: commentId },
+                    {
+                        _id: commentId,
+                        commentAuthor: context.user._id
+                    },
                     { $set: { commentText } },
                     { new: true }
                 );
@@ -182,10 +238,17 @@ const resolvers = {
                 throw new Error(`Error updating comment: ${error.message}`);
             }
         },
-        removeComment: async (parent, { commentId }) => {
+        removeComment: async (parent, { commentId }, context) => {
+            console.log("test removeComment");
+            if (!context.user) {
+                throw new AuthenticationError('You need to be logged in to remove a comment');
+            }
             try {
                 const deletedComment = await Comment.findOneAndDelete(
-                    { _id: commentId }
+                    {
+                        _id: commentId,
+                        commentAuthor: context.user._id,
+                    }
                 );
 
                 if (!deletedComment) {
